Individual has raw fitness instead of normalized
The problem with this change is that it will completely de-generalize my
algorithms, making them only useful for minimization problems (i.e. goal
of 0), and it will mean quite a few tedious comparison fixes. I really
do not think that normalization is what's causing my elitism problem,
but if I do want to try this, I'm stashing it on a feature branch rather
than discarding it.