Implementing raw fitness for Individual
Comparison works as expected: for Individuals that need to be minimized,
the comparison operators return the lesser of the raw fitnesses as the
greater fitness (such that a fitness of 0 is better than any positive
fitness). This also works between two individuals.

Might want to overload for parameter on the LHS as well, just didn't
bother because I have so far only used Individual >|< parameter notation

This will mean readjusting the constant for the SA. Also, at some point
it should be moved to the SA rather than be in the problem.