Revert "Testing change to jumping_mutate()"
This reverts commit bae47e12718e0d11ede0b1f3437054ff4da6396a.

I committed and reverted to preserve the history of the test. Although
it didn't seem like the change would affect the outcome except for a
speedup, it drastically reduced the algorithm's efficiency. The
solutions found ended up being several orders of magnitude worse, even
with additional generations to increase the search space.